SAM2/Search images/requirements.txt
# 基础图像处理库
opencv-python>=4.5.0
# 可选：如果需要使用SURF特征，请安装opencv-contrib-python替代opencv-python
# opencv-contrib-python>=4.5.0
Pillow>=8.0.0
imagehash>=4.2.0
numpy>=1.19.0
scikit-image>=0.18.0
matplotlib>=3.3.0

# PyTorch相关依赖 - 用于ResNet50特征提取
torch>=1.9.0
torchvision>=0.10.0 